+++ src/zwave/interfaces.ts
@@ -53,6 +53,7 @@ export interface IZWaveController extends EventEmitter {
   ): this;
   on(event: 'node added', listener: (node: ZWaveNode) => void): this;
   on(event: 'node ready', listener: (node: ZWaveNode) => void): this;
+  on(event: 'node updated', listener: (node: ZWaveNode) => void): this;
   on(event: 'node removed', listener: (node: ZWaveNode) => void): this;
   on(event: 'value notification', listener: (node: ZWaveNode, args: ZWaveValueEvent) => void): this;
   on(event: 'value updated', listener: (node: ZWaveNode, args: ZWaveValueEvent) => void): this;